[Attempt 2] Added containers for connected components

+++ src/transaction.py
@@ -2,5 +2,5 @@ class MempoolTransaction():
     def __init__(self, txid, fee, weight, parents):
         self.txid = txid
         self.fee = int(fee)
-        self.weight = float(weight)
+        self.weight = int(weight)
         self.parents = parents.split(";") if not parents == "" else []
